backend/requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# CORS and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database and Storage
sqlalchemy==2.0.29
alembic==1.13.1

# Vector Database
weaviate-client==4.5.4

# LLM Providers
openai==1.14.3
anthropic==0.21.3
httpx==0.27.0

# Embeddings and ML
sentence-transformers==2.4.0
transformers==4.36.0
torch==2.1.0
numpy==1.24.4
scikit-learn==1.3.2

# Data Processing
pandas==2.0.3
pydantic==2.5.0
pydantic-settings==2.1.0

# Caching
redis==5.0.3
hiredis==2.3.2

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
tenacity==8.2.3
aiofiles==23.2.1

# Development and Testing
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
black==24.3.0
isort==5.13.2
flake8==7.0.0
mypy==1.9.0 